# backend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "My API"
    },
    "paths": {
...
        "/api/QuanLyDatVe/LayDanhSachPhongVe": {
            "get": {
                "tags": [
                    "QuanLyDatVe"
                ],
                "operationId": "LayDanhSachPhongVe",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaLichChieu",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 0
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/QuanLyNguoiDung/LayDanhSachLoaiNguoiDung": {
            "get": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "LayDanhSachLoaiNguoiDung",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/DangNhap": {
            "post": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "DangNhap",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "ndDN",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/ThongTinDangNhapVM"
                        }
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/api/QuanLyNguoiDung/LayDanhSachNguoiDung": {
            "get": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "LayDanhSachNguoiDung",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tuKhoa",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/LayDanhSachNguoiDungPhanTrang": {
            "get": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "LayDanhSachNguoiDungPhanTrang",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tuKhoa",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "soTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    },
                    {
                        "name": "soPhanTuTrenTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 20
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/TimKiemNguoiDung": {
            "get": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "TimKiemNguoiDung",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tuKhoa",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/TimKiemNguoiDungPhanTrang": {
            "get": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "TimKiemNguoiDungPhanTrang",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "MaNhom",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": "GP01"
                    },
                    {
                        "name": "tuKhoa",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "default": ""
                    },
                    {
                        "name": "soTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    },
                    {
                        "name": "soPhanTuTrenTrang",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "format": "int32",
                        "default": 1
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/ThongTinTaiKhoan": {
            "post": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "ThongTinTaiKhoan",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/LayThongTinNguoiDung": {
            "post": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "LayThongTinNguoiDung",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "taiKhoan",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/ThemNguoiDung": {
            "post": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "ThemNguoiDung",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "nd",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/NguoiDungVM"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/CapNhatThongTinNguoiDung": {
            "put": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "CapNhatThongTinNguoiDung",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "nd",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/NguoiDungVM"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "CapNhat",
                "consumes": [
                    "application/json-patch+json",
                    "application/json",
                    "text/json",
                    "application/*+json"
                ],
                "produces": [],
                "parameters": [
                    {
                        "name": "nd",
                        "in": "body",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/NguoiDungVM"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/QuanLyNguoiDung/XoaNguoiDung": {
            "delete": {
                "tags": [
                    "QuanLyNguoiDung"
                ],
                "operationId": "XoaNguoiDung",
                "consumes": [],
                "produces": [],
                "parameters": [
                    {
                        "name": "TaiKhoan",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Nhập token bearer",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "TokenCybersoft",
                        "in": "header",
                        "description": "Nhập token cybersoft",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ThongTinDangNhapVM": {
            "type": "object",
            "properties": {
                "taiKhoan": {
                    "type": "string"
                },
                "matKhau": {
                    "type": "string"
                }
            }
        },
...
        "NguoiDungVM": {
            "type": "object",
            "properties": {
                "taiKhoan": {
                    "type": "string"
                },
                "matKhau": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "soDt": {
                    "type": "string"
                },
                "maNhom": {
                    "type": "string"
                },
                "maLoaiNguoiDung": {
                    "type": "string"
                },
                "hoTen": {
                    "type": "string"
                }
            }
        },
...
    }
}